+++ lib/sylt_parser.dart
@@ -6,5 +6,6 @@ library;
 export 'src/sylt_lyrics_data.dart';
 export 'src/sylt_lyrics_from_mp3_parser.dart';
 export 'src/enums.dart';
+export 'src/exceptions.dart';
 
 // TODO: Export any libraries intended for clients of this package.