+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Added an exception in case no SYLT-frame can be found in a list of bytes to be parsed.
+- Added documentation for public facing API methods and classes.
+
 ## 0.1.1
 
 - When outputting to SRT format, blank lines are no replaced with a single space character so parsers working with empty lines as separators are less likely to fail, when parsing the output SRT lyrics.